requirements.txt
hydra-core
colorama
func_timeout
tqdm
termcolor
openai
